# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Satta Matka Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdraw-requests:
    post:
      operationId: createWithdrawRequest
      tags: [withdraw]
      summary: Create withdraw request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawRequest"
    get:
      operationId: getWithdrawRequests
      tags: [withdraw]
      summary: Get withdraw requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Withdraw requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WithdrawRequest"
  /withdraw-requests/{id}:
    patch:
      operationId: updateWithdrawRequest
      tags: [withdraw]
      summary: Approve/reject withdraw (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawRequest"

  # Notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WithdrawRequest:
      type: object
      required: [id, userId, amount, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        amount:
          type: number
        bankAccount:
          type: ["string", "null"]
        ifscCode:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    WithdrawRequestInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        bankAccount:
          type: string
        ifscCode:
          type: string
        upiId:
          type: string
        accountName:
          type: string
        bankName:
          type: string
    WithdrawRequestUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        adminNote:
          type: ["string", "null"]

    # Notification
...